Moved some logic into utils file
